requirements.txt
agentscope==1.0.18
pydantic==2.12.5
python-dotenv==1.2.1
requests==2.32.5
beautifulsoup4==4.14.3

# Optional LLM provider adapters (install as needed)
# langchain-ollama==1.0.1
# langchain-openai==1.1.12
# langchain-google-genai==4.2.1
